Fix MSVC warning treated as error in test_server_base_edge_cases.cpp
- Change timedOutClientsCount from std::atomic<int> to std::atomic<size_t>
- Fixes C4267 warning about size_t to int conversion
- Resolves MSVC compilation error on Windows CI

+++ tests/test_server_base_edge_cases.cpp
@@ -54,7 +54,7 @@ class EdgeCaseServer : public ServerBase<EdgeCaseState> {
     std::atomic<int> totalMessagesReceived{0};
     std::atomic<int> totalMessagesSent{0};
     std::atomic<int> errorCount{0};
-    std::atomic<int> timedOutClientsCount{0};
+    std::atomic<size_t> timedOutClientsCount{0};
     std::atomic<int> disconnectCount{0};
 
     Port getActualPort() const {
